YD-87 Extracted base test class
