+++ core/src/test/java/nu/yona/server/goals/service/ActivityCategoryServiceIntegrationTest.java
@@ -2,21 +2,17 @@ package nu.yona.server.goals.service;
 
 import static org.hamcrest.collection.IsIterableContainingInAnyOrder.containsInAnyOrder;
 import static org.junit.Assert.assertThat;
-import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.when;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
-import org.mockito.invocation.InvocationOnMock;
-import org.mockito.stubbing.Answer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.concurrent.ConcurrentMapCache;
@@ -33,41 +29,18 @@ import nu.yona.server.goals.service.ActivityCategoryServiceIntegrationTest.TestC
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = { TestConfiguration.class })
-public class ActivityCategoryServiceIntegrationTest
+public class ActivityCategoryServiceIntegrationTest extends ActivityCategoryServiceTestBase
 {
-	private Set<ActivityCategory> activityCategories = new HashSet<ActivityCategory>();
-
 	@Autowired
 	ActivityCategoryRepository mockRepository;
 
 	@Autowired
 	private ActivityCategoryService service;
 
-	private ActivityCategory gambling;
-	private ActivityCategory news;
-
 	@Before
 	public void setUp()
 	{
-		gambling = ActivityCategory.createInstance("gambling", false, new HashSet<String>(Arrays.asList("poker", "lotto")),
-				Collections.emptySet());
-		news = ActivityCategory.createInstance("news", false, new HashSet<String>(Arrays.asList("refdag", "bbc")),
-				Collections.emptySet());
-
-		activityCategories.add(gambling);
-		activityCategories.add(news);
-
-		when(mockRepository.findAll()).thenReturn(Collections.unmodifiableSet(activityCategories));
-		when(mockRepository.findOne(gambling.getID())).thenReturn(gambling);
-		when(mockRepository.findOne(news.getID())).thenReturn(news);
-		when(mockRepository.save(any(ActivityCategory.class))).thenAnswer(new Answer<ActivityCategory>() {
-			@Override
-			public ActivityCategory answer(InvocationOnMock invocation) throws Throwable
-			{
-				Object[] args = invocation.getArguments();
-				return (ActivityCategory) args[0];
-			}
-		});
+		setUp(mockRepository);
 	}
 
 	private void assertGetAllActivityCategoriesResult(String reason, String... names)
