+++ core/src/test/java/nu/yona/server/goals/service/ActivityCategoryServiceTest.java
@@ -3,14 +3,11 @@ package nu.yona.server.goals.service;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.collection.IsIterableContainingInAnyOrder.containsInAnyOrder;
 import static org.junit.Assert.assertThat;
-import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -21,48 +18,23 @@ import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.invocation.InvocationOnMock;
 import org.mockito.runners.MockitoJUnitRunner;
-import org.mockito.stubbing.Answer;
 
 import nu.yona.server.goals.entities.ActivityCategory;
 import nu.yona.server.goals.entities.ActivityCategoryRepository;
 
 @RunWith(MockitoJUnitRunner.class)
-public class ActivityCategoryServiceTest
+public class ActivityCategoryServiceTest extends ActivityCategoryServiceTestBase
 {
-	private Set<ActivityCategory> activityCategories = new HashSet<ActivityCategory>();
-
 	@Mock
 	ActivityCategoryRepository mockRepository = mock(ActivityCategoryRepository.class);
 	@InjectMocks
 	private ActivityCategoryService service = new ActivityCategoryService();
 
-	private ActivityCategory gambling;
-	private ActivityCategory news;
-
 	@Before
 	public void setUp()
 	{
-		gambling = ActivityCategory.createInstance("gambling", false, new HashSet<String>(Arrays.asList("poker", "lotto")),
-				Collections.emptySet());
-		news = ActivityCategory.createInstance("news", false, new HashSet<String>(Arrays.asList("refdag", "bbc")),
-				Collections.emptySet());
-
-		activityCategories.add(gambling);
-		activityCategories.add(news);
-
-		when(mockRepository.findAll()).thenReturn(Collections.unmodifiableSet(activityCategories));
-		when(mockRepository.findOne(gambling.getID())).thenReturn(gambling);
-		when(mockRepository.findOne(news.getID())).thenReturn(news);
-		when(mockRepository.save(any(ActivityCategory.class))).thenAnswer(new Answer<ActivityCategory>() {
-			@Override
-			public ActivityCategory answer(InvocationOnMock invocation) throws Throwable
-			{
-				Object[] args = invocation.getArguments();
-				return (ActivityCategory) args[0];
-			}
-		});
+		setUp(mockRepository);
 	}
 
 	/*
